# openapi-server/src/api_default.go
/*
 * Title
 *
 * Title
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package mzserver

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/go-chi/chi/v5"
)

// DefaultApiController binds http requests to an api service and writes the service results to the http response
type DefaultApiController struct {
	service DefaultApiServicer
	errorHandler ErrorHandler
}

// DefaultApiOption for how the controller is set up.
type DefaultApiOption func(*DefaultApiController)

// WithDefaultApiErrorHandler inject ErrorHandler into controller
func WithDefaultApiErrorHandler(h ErrorHandler) DefaultApiOption {
	return func(c *DefaultApiController) {
		c.errorHandler = h
	}
}

// NewDefaultApiController creates a default api controller
func NewDefaultApiController(s DefaultApiServicer, opts ...DefaultApiOption) Router {
	controller := &DefaultApiController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the DefaultApiController
func (c *DefaultApiController) Routes() Routes {
	return Routes{ 
		{
			"Get3dMap",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/3dmap",
			c.Get3dMap,
		},
		{
			"GetChromatogramCount",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/chromatograms/count",
			c.GetChromatogramCount,
		},
		{
			"GetChromatogramData",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/chromatograms/{chromatogramId}",
			c.GetChromatogramData,
		},
		{
			"GetChromatogramHeader",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/chromatograms/{chromatogramId}/header",
			c.GetChromatogramHeader,
		},
		{
			"GetChromatograms",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/chromatograms",
			c.GetChromatograms,
		},
		{
			"GetFileName",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/filename",
			c.GetFileName,
		},
		{
			"GetInstrumentAnalyzer",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/instrument/analyzer",
			c.GetInstrumentAnalyzer,
		},
		{
			"GetInstrumentData",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/instrument",
			c.GetInstrumentData,
		},
		{
			"GetInstrumentDetector",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/instrument/detector",
			c.GetInstrumentDetector,
		},
		{
			"GetInstrumentManufacturer",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/instrument/manufacturer",
			c.GetInstrumentManufacturer,
		},
		{
			"GetInstrumentModel",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/instrument/model",
			c.GetInstrumentModel,
		},
		{
			"GetIonisationMethod",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/instrument/ionisation",
			c.GetIonisationMethod,
		},
		{
			"GetIsolationWindows",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/IsolationWindow",
			c.GetIsolationWindows,
		},
		{
			"GetSampleData",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/instrument/sample",
			c.GetSampleData,
		},
		{
			"GetScanCount",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/scans/count",
			c.GetScanCount,
		},
		{
			"GetScanHeader",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/scans/{id}/header",
			c.GetScanHeader,
		},
		{
			"GetScanPeakCount",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/scans/{id}/peaks/count",
			c.GetScanPeakCount,
		},
		{
			"GetScanPeaks",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/scans/{id}/peaks",
			c.GetScanPeaks,
		},
		{
			"GetScansData",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/scans",
			c.GetScansData,
		},
		{
			"GetSoftware",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/instrument/software",
			c.GetSoftware,
		},
		{
			"GetSource",
			strings.ToUpper("Get"),
			"/msdata/{msDataId}/instrument/source",
			c.GetSource,
		},
		{
			"PostFile",
			strings.ToUpper("Post"),
			"/file",
			c.PostFile,
		},
	}
}

// Get3dMap - 
func (c *DefaultApiController) Get3dMap(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.Get3dMap(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetChromatogramCount - 
func (c *DefaultApiController) GetChromatogramCount(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetChromatogramCount(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetChromatogramData - 
func (c *DefaultApiController) GetChromatogramData(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	chromatogramIdParam, err := parseInt32Parameter(chi.URLParam(r, "chromatogramId"), true)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}

	result, err := c.service.GetChromatogramData(r.Context(), msDataIdParam, chromatogramIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetChromatogramHeader - 
func (c *DefaultApiController) GetChromatogramHeader(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	chromatogramIdParam, err := parseInt32Parameter(chi.URLParam(r, "chromatogramId"), true)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}

	result, err := c.service.GetChromatogramHeader(r.Context(), msDataIdParam, chromatogramIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetChromatograms - 
func (c *DefaultApiController) GetChromatograms(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetChromatograms(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetFileName - 
func (c *DefaultApiController) GetFileName(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetFileName(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetInstrumentAnalyzer - 
func (c *DefaultApiController) GetInstrumentAnalyzer(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetInstrumentAnalyzer(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetInstrumentData - 
func (c *DefaultApiController) GetInstrumentData(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetInstrumentData(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetInstrumentDetector - 
func (c *DefaultApiController) GetInstrumentDetector(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetInstrumentDetector(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetInstrumentManufacturer - 
func (c *DefaultApiController) GetInstrumentManufacturer(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetInstrumentManufacturer(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetInstrumentModel - 
func (c *DefaultApiController) GetInstrumentModel(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetInstrumentModel(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetIonisationMethod - 
func (c *DefaultApiController) GetIonisationMethod(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetIonisationMethod(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetIsolationWindows - 
func (c *DefaultApiController) GetIsolationWindows(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetIsolationWindows(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetSampleData - 
func (c *DefaultApiController) GetSampleData(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetSampleData(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetScanCount - 
func (c *DefaultApiController) GetScanCount(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetScanCount(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetScanHeader - 
func (c *DefaultApiController) GetScanHeader(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetScanHeader(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetScanPeakCount - 
func (c *DefaultApiController) GetScanPeakCount(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetScanPeakCount(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetScanPeaks - 
func (c *DefaultApiController) GetScanPeaks(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetScanPeaks(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetScansData - 
func (c *DefaultApiController) GetScansData(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetScansData(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetSoftware - 
func (c *DefaultApiController) GetSoftware(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetSoftware(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// GetSource - 
func (c *DefaultApiController) GetSource(w http.ResponseWriter, r *http.Request) {
	msDataIdParam := chi.URLParam(r, "msDataId")
	
	result, err := c.service.GetSource(r.Context(), msDataIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}

// PostFile - 
func (c *DefaultApiController) PostFile(w http.ResponseWriter, r *http.Request) {
	bodyParam := *os.File{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&bodyParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	result, err := c.service.PostFile(r.Context(), bodyParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, result.Headers, w)

}